Add files via upload
Student marks prediction is a popular data science case study based on the problem of regression. It is a good regression problem for data science beginners as it is easy to solve and understand. So if you want to learn how to predict the marks of a student with machine learning, this article is for you. In this article, I will take you through the task of student marks prediction with machine learning using Python.

Student Marks Prediction (Case Study)
You are given some information about students like:

the number of courses they have opted for
the average time studied per day by students
marks obtained by students